mark-flow-ui: Extract PartySelectionPage from VxMark and VxMarkScan
Mirrors the StartPage / ContestPage / ReviewPage pattern: a
presentational component that takes election data, vote state, and
URL strings as props. The thin wrappers in apps/mark and apps/mark-scan
read from BallotContext and pass through.

Removes the byte-for-byte duplication of party_selection_screen.tsx
across the two apps that landed in the previous commit.
